backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
